+++ src/sys/gl.rs
@@ -119,6 +119,7 @@ pub trait Gl: protected::Sealed {
     location: &Self::UniformLocation,
   ) -> [f32; N];
 
+  fn set_uniform_1f(&self, location: &Self::UniformLocation, data: f32);
   fn set_uniform_1i(&self, location: &Self::UniformLocation, data: i32);
   fn set_uniform_1ui(&self, location: &Self::UniformLocation, data: u32);
   fn set_uniform_1iv(&self, location: &Self::UniformLocation, data: &[i32]);